+++ core/dbt/include/global_project/macros/schema_tests/accepted_values.sql
@@ -31,7 +31,7 @@ validation_errors as (
     )
 )
 
-select count(*) as validation_errors
+select *
 from validation_errors
 
 {% endmacro %}