+++ test/integration/005_simple_seed_test/macros/schema_test.sql
@@ -1,5 +1,5 @@
 
-{% macro test_column_type(model, column_name, type) %}
+{% test column_type(model, column_name, type) %}
 
     {% set cols = adapter.get_columns_in_relation(model) %}
 
@@ -8,12 +8,15 @@
         {% do col_types.update({col.name: col.data_type}) %}
     {% endfor %}
 
+    {% set validation_message = 'Got a column type of ' ~ col_types.get(column_name) ~ ', expected ' ~ type %}
+
     {% set val = 0 if col_types.get(column_name) == type else 1 %}
     {% if val == 1 and execute %}
-        {# I'm so tired of guessing what's wrong, let's just log it #}
-        {{ log('Got a column type of ' ~ col_types.get(column_name) ~ ', expected ' ~ type, info=True) }}
+        {{ log(validation_message, info=True) }}
     {% endif %}
 
-    select {{ val }} as pass_fail
+    select '{{ validation_message }}' as validation_error
+    from (select true) as nothing
+    where {{ val }} = 1
 
-{% endmacro %}
+{% endtest %}