+++ test/integration/008_schema_tests_test/macros-v2/macros/tests.sql
@@ -1,21 +1,15 @@
+{% test every_value_is_blue(model, column_name) %}
 
-
-{% macro test_every_value_is_blue(model, column_name) %}
-
-    select
-        count(*)
-
+    select *
     from {{ model }}
     where {{ column_name }} != 'blue'
 
-{% endmacro %}
+{% endtest %}
 
 
-{% macro test_rejected_values(model, column_name, values) %}
+{% test rejected_values(model, column_name, values) %}
 
-    select
-        count(*)
-
+    select *
     from {{ model }}
     where {{ column_name }} in (
         {% for value in values %}
@@ -23,20 +17,19 @@
         {% endfor %}
     )
 
-{% endmacro %}
+{% endtest %}
 
 
-{% macro test_equivalent(model, value) %}
+{% test equivalent(model, value) %}
     {% set expected = 'foo-bar' %}
-    select
-    {% if value == expected %}
-        0
-    {% else %}
-        {% set msg -%}
-        got "{{ value }}", expected "{{ expected }}"
-        {%- endset %}
-        {% do log(msg, info=True) %}
-        1
+    {% set eq = 1 if value == expected else 0 %}
+    {% set validation_message -%}
+      'got "{{ value }}", expected "{{ expected }}"'
+    {%- endset %}
+    {% if eq == 0 and execute %}
+        {{ log(validation_message, info=True) }}
     {% endif %}
-    as id
-{% endmacro %}
+
+    select {{ validation_message }} as validation_error
+    where {{ eq }} = 0
+{% endtest %}