+++ test/integration/008_schema_tests_test/test-context-macros/custom_schema_tests.sql
@@ -1,19 +1,19 @@
-{% macro test_type_one(model) %}
+{% test type_one(model) %}
 
-    select count(*) from (
+    select * from (
 
         select * from {{ model }}
         union all
         select * from {{ ref('model_b') }}
-        
+
     ) as Foo
 
-{% endmacro %}
+{% endtest %}
 
-{% macro test_type_two(model) %}
+{% test type_two(model) %}
 
     {{ config(severity = "WARN") }}
 
-    select count(*) from {{ model }}
+    select * from {{ model }}
 
-{% endmacro %}
+{% endtest %}