+++ test/integration/008_schema_tests_test/test_schema_v2_tests.py
@@ -126,52 +126,6 @@ class TestMalformedSchemaTests(DBTIntegrationTest):
             self.run_dbt(strict=False)
 
 
-class TestMalformedMacroTests(DBTIntegrationTest):
-
-    def setUp(self):
-        DBTIntegrationTest.setUp(self)
-        self.run_sql_file("seed.sql")
-
-    @property
-    def schema(self):
-        return "schema_tests_008"
-
-    @property
-    def models(self):
-        return "models-v2/custom-bad-test-macro"
-
-    @property
-    def project_config(self):
-        return {
-            'config-version': 2,
-            "macro-paths": ["macros-v2/malformed"],
-        }
-
-    def run_schema_validations(self):
-        args = FakeArgs()
-        test_task = TestTask(args, self.config)
-        return test_task.run()
-
-    @use_profile('postgres')
-    def test_postgres_malformed_macro_reports_error(self):
-        self.run_dbt(['deps'])
-        self.run_dbt(strict=True)
-        expected_failure = 'not_null'
-
-        test_results = self.run_schema_validations()
-
-        self.assertEqual(len(test_results), 2)
-
-        for result in test_results:
-            self.assertIn(result.status, ('error', 'fail'))
-            # Assert that error is thrown for empty schema test
-            if result.status == "error":
-                self.assertIn("Returned 0 rows", result.message)
-            # Assert that failure occurs for normal schema test
-            elif result.status == "fail":
-                self.assertIn(expected_failure, result.node.name)
-
-
 class TestHooksInTests(DBTIntegrationTest):
 
     @property
@@ -220,9 +174,6 @@ class TestCustomSchemaTests(DBTIntegrationTest):
         return {
             'packages': [
                 {
-                    "local": "./local_dependency",
-                },
-                {
                     'git': 'https://github.com/fishtown-analytics/dbt-integration-project',
                     'revision': 'dbt/0.17.0',
                 },
@@ -256,7 +207,7 @@ class TestCustomSchemaTests(DBTIntegrationTest):
         self.assertEqual(len(results), 4)
 
         test_results = self.run_schema_validations()
-        self.assertEqual(len(test_results), 6)
+        self.assertEqual(len(test_results), 5)
 
         expected_failures = ['unique', 'every_value_is_blue']