+++ test/integration/022_bigquery_test/macros/partition_metadata.sql
@@ -1,22 +1,19 @@
 
-{% macro test_number_partitions(model, expected) %}
+{% test number_partitions(model, expected) %}
 
     {%- set result = get_partitions_metadata(model) %}
-    
+
     {% if result %}
         {% set partitions = result.columns['partition_id'].values() %}
     {% else %}
         {% set partitions = () %}
     {% endif %}
-        
-    {% set actual = partitions | length %}
-    
-    {% if model and actual == expected %}
-        select 0 as success
-    {% else %}
-        -- actual: {{ actual }}
-        -- expected: {{ expected }}
-        select 1 as error
-    {% endif %}
 
-{% endmacro %}
+    {% set actual = partitions | length %}
+    {% set success = 1 if model and actual == expected else 0 %}
+
+    select 'Expected {{ expected }}, but got {{ actual }}' as validation_error
+    from (select true)
+    where {{ success }} = 0
+
+{% endtest %}