+++ test/integration/041_presto_test/models/table_model.sql
@@ -1,4 +1,0 @@
-
-{{ config(materialized = "table") }}
-
-select * from {{ ref('ephemeral_model') }}